# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FinaMantic",
        "description": "FinaMantic"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/notification": {
            "get": {
                "tags": ["Profile"],
                "summary": "Notifications",
                "description": "Notifications",
                "operationId": "notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all notifications",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/plaid/set_access_token":{
            "post": {
                "tags": [
                    "Plaid"
                ],
                "summary": "Set access token",
                "description": "Get access token by public token",
                "operationId": "set_access_token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/accessToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create link token successfully",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/plaid/transactions":{
            "get": {
                "tags": [
                    "Plaid"
                ],
                "summary": "transactions get",
                "description": "transactions get",
                "operationId": "transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    },
                    {
                        "in": "query",
                        "name": "access_token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "transactions get successfully",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
        "/plaid/identity":{
            "get": {
                "tags": [
                    "Plaid"
                ],
                "summary": "identity get",
                "description": "identity get",
                "operationId": "identity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    },
                    {
                        "in": "query",
                        "name": "access_token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "identity get successfully",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
        "/plaid/balance":{
            "get": {
                "tags": [
                    "Plaid"
                ],
                "summary": "balance get",
                "description": "balance get",
                "operationId": "balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    },
                    {
                        "in": "query",
                        "name": "access_token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "balance get successfully",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/plaid/accounts":{
            "get": {
                "tags": [
                    "Plaid"
                ],
                "summary": "accounts get",
                "description": "accounts get",
                "operationId": "accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    },
                    {
                        "in": "query",
                        "name": "access_token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "accounts get successfully",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "responsesOk": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "responsesError": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "error": {
                    "type": "object"
                }
            }
        },
...
        "accessToken":{
            "type": "object",
            "required": [
                "public_token"
            ],
            "properties": {
                "public_token": {
                    "type": "string",
                    "example": "62046cb11fd9e11c2cbad810"
                }
            }
        }
    },
    "parameters": {
        "HeaderVersion": {
            "name": "X-FinaMantic-Version",
            "in": "header",
            "description": "X-FinaMantic-Version",
            "required": true,
            "type": "string",
            "default": "1.0.0"
        },
        "HeaderLanguage": {
            "name": "Accept-Language",
            "in": "header",
            "description": "Accept-Language",
            "required": true,
            "type": "string",
            "default": "en"
        },
        "HeaderPlatform": {
            "name": "X-FinaMantic-Platform",
            "in": "header",
            "description": "android or ios or web",
            "required": true,
            "type": "string",
            "default": "android"
        },
        "HeaderAuth": {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
        },
...
    }
}